Update requirements to use elephant 0.6.3
Because this resolves the version conflict between Neo and elephant, we can finally include formal requirements in ``setup.py`` using ``install_requires``. This means the package will now install its own dependencies (except for PyAV). This allows the installation instructions to be simplified, as well as the Travis CI configuration.