RAUC disabled by default
Previously, RAUCs were always calculated if ``lazy=False``, i.e., there was no way to disable them if other non-lazy features were needed. Now, RAUCs are not calculated if the bin duration is set to None, which is the new default. So now to calculate RAUCs, ``rauc_bin_duration`` must be given explicitly (as well as ``lazy=False``).