Allow alternative, null values for file, dataset in global config file
Users may expect they could use "false" or "none" in the global config file as a null value to set the file and dataset parameters to their factory defaults. The TOML parser cannot parse a parameter with the result of returning Python's None (the parameter is simply deleted instead), but it can parse "false" to Python's False, so here we allow that to represent a null value. Some code is simplified by using False rather than None in the factory default as well.