Show epoch encoder even when no labels are given (as long as a file is)
Closes #317

Previously, if `epoch_encoder_possible_labels` was not specified, and if an empty file was provided for `epoch_encoder_file`, the epoch encoder would not load at all. A minimum of one possible label needed to be provided in metadata.

With this change, no possible labels need be specified in metadata for the epoch encoder to load. Using the epoch encoder's "New label" button (ephyviewer >= 1.4.0), possible labels can be created entirely on the fly.